convert a number to binary draft

+++ hw-1/epsilon.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    long double epsilon = 1;
+    float epsilon = 1;
     int n = 0;
     while ((1+epsilon) != 1) {
         epsilon /= 10;
