+++ hw-1/float64.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    double epsilon = 1;
+    long double epsilon = 1;
     int n = 0;
     while (1 + epsilon != 1) {
         n++;
